chore: Update fluent-bit config (#108)
This commit modifies the fluent-bit configuration to
include an additional port (8888) for receiving CloudEvents
from Tekton Pipelines. This allows fluent-bit to capture
both Kubernetes events and CloudEvents, providing a more
comprehensive logging solution. The configuration also
includes the necessary inputs and outputs for forwarding
the logs to an OpenSearch cluster with specific formatting
and TLS settings.

Signed-off-by: Sergiy Kulanov <sergiy_kulanov@epam.com>